Check for supercritical Gros Michel going extinct first

+++ objects/stickers/supercritical.lua
@@ -15,7 +15,7 @@ SMODS.Sticker {
         ((supercritical_roll > 0.4) and (supercritical_roll <= 0.7)) and (area == G.shop_jokers or area == G.pack_cards)
   end,
   calculate = function(self, card, context)
-    if context.end_of_round and not context.repetition and not context.individual then
+    if context.end_of_round and not context.repetition and not context.individual and card.config.center_key ~= 'j_gros_michel' or (card.config.center_key == 'j_gros_michel' and context.gros_michel_safe and context.gros_michel_card == card) then
       local last_factor = card.ability.extra.current_factor;
 
       card.ability.extra.supercritical_stage = card.ability.extra.supercritical_stage + 1
@@ -39,13 +39,13 @@ SMODS.Sticker {
                 { message = localize('k_supercritical_reset'), colour = G.C.FILTER, delay = 0.45 })
             end
 
-            if left_joker and not left_joker.ability.eternal then
+            if left_joker and not left_joker.destroyed and not left_joker.ability.eternal then
               SMODS.destroy_cards(left_joker, nil, nil)
               card_eval_status_text(left_joker, 'extra', nil, nil, nil,
                 { message = localize('k_destroyed_ex'), colour = G.C.FILTER, delay = 0.45 })
             end
 
-            if right_joker and not right_joker.ability.eternal then
+            if right_joker and not right_joker.destroyed and not right_joker.ability.eternal then
               SMODS.destroy_cards(right_joker, nil, nil)
               card_eval_status_text(right_joker, 'extra', nil, nil, nil,
                 { message = localize('k_destroyed_ex'), colour = G.C.FILTER, delay = 0.45 })
